Expresions 100% Assignments 50%
Everything is pain and im already DEAD

+++ test2.fs
@@ -1,12 +1,13 @@
 var a:int = 10
+
 var b:int = 4
 
 var c:int = a * b
 
 var d:int = d / c + (b * a)
 
-var e:float = 3.000 / d
+var e:double = 3.000 / d
 
-var f:double = 400d
+var f:double = 400
